Cleaning things up to reduce memory footprint
Previously was storing a lot of data for convenience purposes but now we're just storing a 2d array of 1/0 (alive/dead) and we'll work from that. Further improvements could be made if we want to keep track of living cells instead of all cells but for now we'll just accept that we only track the board at any given time.

Also changed how we're drawing the board to reduce complexity of previous method. The issue here is that spacing isn't quite right when we resize the grid so changes are needed to fix that.